feat: start the events recommendations

+++ db/migrate/20231225123452_create_eras.rb
@@ -3,7 +3,7 @@ class CreateEras < ActiveRecord::Migration[7.0]
     create_table :eras do |t|
       t.string :name
       t.integer :tier
-      t.float :point, default: 0.0
+      t.integer :point, default: 0
 
       t.timestamps
     end
